Relocating `AreaConfig` class
